==> 1_scripts/6_dimred_clustering.Rmd <==
---
title: "Dimensionality reduction and clustering"
author: "*Lisa Loksø Dietz*"
date: "*`r format(Sys.time(), '%Y-%m-%d')`*"
knit: (function(inputFile, encoding) { 
      out_dir = paste0(dirname(getwd()),"/3_output");
      rmarkdown::render(inputFile,
                        encoding=encoding, 
                        output_file=file.path(out_dir, '5_dimred_clustering.html')) })
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 4
    toc_collapsed: true
    code_folding: show
---

```{r Setup 1, include = FALSE}

knitr::opts_chunk$set(fig.width=10, fig.height=7, out.width = "70%",fig.align = 'center', eval=TRUE) 
# eval controls whether code should be run when knitting
# cache will avoid running again the code chunk again if you didn't change anything within it

```

## 0) Load source and set dir

```{r Directories, message = FALSE, warning = FALSE}

rm(list=ls())
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))
source("0_source.R")

```

## 1) Load data

### 1.1) Cleaned flowSet

```{r Read files, message = FALSE}

read_dir = glue("{dir_project}/2_pipeline/{folder_QC}/out")
fs_clean = readRDS(glue("{read_dir}/FlowSet_clean.rds"))
pData(fs_clean)

```

### 1.2) Panel

```{r Read panel, message = FALSE}

# read panel
panel = read.csv(file = glue("{dir_project}/2_pipeline/{folder_QC}/out/Panel.csv"))[,-1]

# define type and state markers
gated_markers = c("CD14_16","CD19","CD3")
features_type = setdiff(panel[panel$marker_class=="type","antigen"], gated_markers)
features_state= panel[panel$marker_class=="state","antigen"]
features_inhibitory = c("TIGIT","PD1","CD39")
features_chemokineR = c("CCR4","CXCR5")
features_cytokine_degran = c("CD107a","IFN","TNF","IL2","IL413")
features_effectorproteins = c("Perforin","GzmB","GzmK","Granulysin")
features_TF = c("Tbet","TCF1")
features_prolif = c("Ki67")
features_clustering = features_type

# choose channels to use in SCE
panel[panel$fcs_colname == "SSC-A", "antigen"] = "SSC"
panel[panel$fcs_colname == "SSC-A", "marker_class"] = "type"
channels_sce = panel[(panel$marker_class!="none")&(!panel$antigen %in% gated_markers),"fcs_colname"]

total_markers = length(channels_sce)

```

## 2) Clustering and dimensionality reduction

### 2.1) Create Single Cell Experiment (SCE)

```{r Create SCE, message = FALSE}

set.seed(1234)
sce = prepData(fs_clean,
               md = pData(fs_clean),
               panel = panel,
               features = channels_sce,
               FACS = TRUE,
               transform = FALSE,
               panel_cols = list(channel = "fcs_colname",
                                 antigen = "antigen"
                                 ),
               md_cols = list(file = "name",
                              id = "name",
                              factors = c("stimulation"
                                          )
                              )
               )

assayNames(sce)[1] = "exprs"

saveRDS(sce, file = glue("{dir_store}/sce.rds"))
sce = readRDS(file = glue("{dir_store}/sce.rds"))


rowData(sce)
colData(sce)
sce


```

### 2.2) Cluster SCE and delta area plot

```{r Cluster and delta area plot, message = FALSE}

set.seed(1234)
no_clusters = 25
meta_name = glue("meta{no_clusters}")
features_clustering

sce = cluster(sce,
              features = features_clustering,
              maxK = no_clusters,
              xdim = 10,
              ydim = 10,
              verbose = TRUE,
              seed = 1234)

delta_area(sce)

```

### 2.3 Save and read clustered SCE

```{r Save clustered data, message = FALSE}

saveRDS(sce, file = glue("{dir_store}/sce_clust.rds"))
sce = readRDS(file = glue("{dir_store}/sce_clust.rds"))
sce

```

### 2.4) Dimensionality reduction

```{r Run dimensionality reduction, message = FALSE}

set.seed(1234)
sce = runDR(sce,
            "UMAP",
            cells = NULL,
            features = features_clustering)

```

### 2.5) Save and read dimred SCE

```{r Save SCE, message = FALSE}

saveRDS(sce, glue("{dir_out}/sce_clust_dimred.rds"))
sce = readRDS(file = glue("{dir_out}/sce_clust_dimred.rds"))
sce$stimulation = factor(sce$stimulation,levels = c("NEG", "GAG", "SEB"))

```

## 3) UMAP visualisations

### 3.1) FlowSOM clusters

```{r Visualize FlowSOM clusters on UMAP, message = FALSE, fig.width=14, fig.height=12}

width = 14
height = 12

width = 14*0.5
height = 12*0.5

png(file = glue("{dir_store}/UMAP_meta25.png"), width=width,height=height,units="in",res=200)
q_plotDR(sce,
         color_by = "meta25",
         labelsize = 5)
dev.off()

png(file = glue("{dir_store}/UMAP_facet_stim_meta25.png"), width=24,height=height,units="in",res=200)
q_plotDR(sce,
         color_by = "meta25",
         labelsize = 5,
         facet_by = "stimulation")
dev.off()

png(file = glue("{dir_store}/UMAP_clust.png"), width=12,height=5,units="in",res=200)
plotDR(sce,
       color_by = features_clustering,
       ncol = 4)
dev.off()

width = 5
height = 4.29


for(feature in features_clustering){
  png(file = glue("{dir_store}/UMAP_{feature}.png"), width=width,height=height,units="in",res=200)
  print(plotDR(sce,
         color_by = feature))
  dev.off()
}

plot_list = list()
for(feature in c(features_type,features_state)){
  p = plotDR(sce,
         color_by = feature,
         facet_by = "stimulation")
  plot_list[[feature]] = p
}

png(file = glue("{dir_store}/UMAP_facet_stim_all.png"), width=14,height=100,units="in",res=200)
grid.arrange(grobs = plot_list, ncol = 1)
dev.off()

for(feature in c(features_type,features_state)){
  png(file = glue("{dir_store}/UMAP_facet_stim_{feature}.png"), width=14,height=12,units="in",res=200)
  print(plotDR(sce,
         color_by = feature,
         facet_by = "stimulation"))
  dev.off()
}



png(file = glue("{dir_store}/density_meta25.png"), width = 16, height = 14, units = "in", res=200)
plotClusterExprs(sce,
                 k = "meta25",
                 features = features_clustering)
dev.off()


```

### 3.2) Manual cell types

#### 3.2.1) Import aggregate labels

```{r Load SCE, message = FALSE}

aggregate_labels = readRDS(glue("{dir_project}/2_pipeline/{folder_celltype}/out/aggregate_labels.rds"))
manual_labels_SCE = as.factor(aggregate_labels)
sce$manual_id = manual_labels_SCE

```

#### 3.2.2) Plot labels on UMAP

```{r UMAP, message = F, fig.width = 20, fig.height = 20}

cell_types_of_interest <- setdiff(levels(sce$manual_id),"Unlabeled")
width = 14*0.8
height = 12*0.8
png(filename = glue("{dir_store}/UMAP_manual.png"),width = width, height = height, units = "in", res = 200)
q_plotDR(sce,
         color_by = "manual_id",
         labelsize = 5,
         highlight_clust = cell_types_of_interest)
dev.off()

png(filename = glue("{dir_store}/UMAP_manual_facet_stim.png"),width = 24, height = 12, units = "in", res = 200)
q_plotDR(sce,
         color_by = "manual_id",
         labelsize = 5,
         highlight_clust = cell_types_of_interest,
         facet_by = "stimulation")
dev.off()

cell_types_of_interest <- setdiff(levels(sce$manual_id),"Unlabeled")
png(filename = glue("{dir_store}/UMAP_manual.png"),width = 14, height = 12, units = "in", res = 200)
q_plotDR(sce,
         color_by = "manual_id",
         labelsize = 5,
         highlight_clust = cell_types_of_interest)
dev.off()

```

### 3.3) Marker positivity

#### 3.3.1) Import boolean matrix

```{r import matrix, message=F}

boolean_matrix_marker_positivity = read.csv(file = glue("{dir_project}/2_pipeline/{folder_marker}/out/boolean_matrix_marker+.csv"))[,-1]
names(boolean_matrix_marker_positivity) = gsub(".$","_pos",names(boolean_matrix_marker_positivity))
marker_positivity_to_plot = names(boolean_matrix_marker_positivity)[-1]

```

#### 3.3.2) Add to colData(sce)

```{r add to coldata, message = F}

new_coldata = cbind(colData(sce),boolean_matrix_marker_positivity)
sce = scdrake::sce_add_colData(sce, new_coldata, replace = TRUE)

image(as.matrix(boolean_matrix_marker_positivity))
colnames(boolean_matrix_marker_positivity)

```

#### 3.3.1) Plot marker positivity

```{r}

plot_list = list()
for(color_plot in marker_positivity_to_plot){
  p = plotDR(sce,
       color_by = color_plot,
       facet_by = "stimulation")+
  scale_color_manual(values = c("FALSE" = "grey70", "TRUE" = "#00FF00"))
  plot_list[[color_plot]] = p
}

png(file = glue("{dir_store}/UMAP_facet_stim_all_pos.png"), width=14,height=100,units="in",res=200)
grid.arrange(grobs = plot_list, ncol = 1)
dev.off()


for(color_plot in marker_positivity_to_plot){
  
  png(filename = glue("{dir_store}/UMAP_facet_stim_{color_plot}.png"), width=14,height=12,units="in",res=200)
  print(plotDR(sce,
       color_by = color_plot,
       facet_by = "stimulation")+
  scale_color_manual(values = c("FALSE" = "grey70", "TRUE" = "#00FF00")))
  dev.off()
  
}

```

#### 3.3.2) Weighted marker positivity (background subtracted)

```{r Weighted positivity, message=FALSE}

colData(sce)$meta25 = cluster_ids(sce,"meta25")

dataframe = q_ICSsfcm_GatingMarkers_BackgroundSub_BulkClust(sce,
                                                            k = "meta25",
                                                            condition_col = "stimulation",
                                                            background_value = "NEG")

dataframe$marker = gsub("_pos","",dataframe$marker)

df_plot =  dataframe[(dataframe$stimulation=="GAG"),c("meta25","marker","n_pos_backgroundsubtracted")]

require(plyr)

n_order = ddply(.data=df_plot, 
                 .(marker), 
                 summarize, 
                 n=round(sum(n_pos_backgroundsubtracted))) %>% arrange(desc(n))

marker_order = n_order$marker

df_plot$marker = factor(df_plot$marker, levels = marker_order)

n_labels <- ddply(.data=df_plot, 
                 .(marker), 
                 summarize, 
                 n=paste("n =", round(sum(n_pos_backgroundsubtracted))))

png(filename = glue("{dir_project}/3_results/clusterdistribution_marker_pies.png"), height = 15,width = 15,unit = "in",res=300)
ggplot(df_plot, aes(x="", y=n_pos_backgroundsubtracted, fill=meta25))+
  geom_bar(width = 1, stat = "identity")+
  facet_wrap(~marker,scales="free")+
  geom_text(data=n_labels,inherit.aes=FALSE, parse=FALSE,aes(x=-.3, y=0,label = n))+
  scale_fill_manual(values = CATALYST:::.cluster_cols[1:25])+coord_polar("y",start=0)+
  theme_void()+
  ggtitle(label = "Distribution of FlowSOM clusters among positive cells")+
  ggeasy::easy_center_title()
dev.off()



```

```{r Project description figure, message = F}

cell_types_of_interest <- setdiff(levels(sce$manual_id),"Unlabeled")

fig1a = q_plotDR(sce,
                 color_by = "manual_id",
                 labelsize = 3,
                 highlight_clust = cell_types_of_interest)+ 
  guides(color=guide_legend(title="Manual gating",override.aes = list(size = 4)))


fig1b = q_plotDR(sce,
                 color_by = "meta25",
                 labelsize = 3)+ 
  guides(color=guide_legend(title="FlowSOM metacluster",override.aes = list(size = 4)))



fig1c = plotDR(sce,
               color_by = features_clustering,
               ncol = 2)

fig1d = q_plotDR_shuffled(filterSCE(sce,
                         stimulation != "SEB"),
               color_by = "IFN_pos",
               facet_by = "stimulation",
               ncol = 1)+
  scale_color_manual(values = c("FALSE" = "grey70", "TRUE" = "red"))+
  guides(color=guide_legend(title="IFN positive",override.aes = list(size = 4)))


df_plot2 =  dataframe[(dataframe$marker %in%c("TNF","IFN","CD107a")&dataframe$stimulation=="GAG"),c("meta25","marker","n_pos_backgroundsubtracted")]
df_plot2$marker = factor(df_plot2$marker,levels = c("IFN","TNF","CD107a"))
n_labels <- ddply(.data=df_plot2, 
                 .(marker), 
                 summarize, 
                 n=paste("n =", round(sum(n_pos_backgroundsubtracted))))
fig1e = print(ggplot(df_plot2, aes(x="", 
                                   y=n_pos_backgroundsubtracted, 
                                   fill=meta25))+
                geom_bar(width = 1, 
                         stat = "identity")+
  facet_wrap(~marker,
             scales="free",
             ncol=1)+
  geom_text(data=n_labels,
            inherit.aes=FALSE, 
            parse=FALSE,
            aes(x=-.3, y=0,label = n))+
  scale_fill_manual(values = CATALYST:::.cluster_cols[1:25])+
    coord_polar("y",start=0)+
  theme_void() +
  easy_remove_legend()
)

fig1a$layers[[1]]$aes_params$size = 0.005
fig1b$layers[[1]]$aes_params$size = 0.005
fig1c$layers[[1]]$aes_params$size = 0.000000001
fig1d$layers[[1]]$aes_params$size = 0.000000001


design <- "
AAAAABBBBB
AAAAABBBBB
AAAAABBBBB
CCCCCDDDEE
CCCCCDDDEE
CCCCCDDDEE
CCCCCDDDEE
"

png(filename = glue("{dir_project}/3_results/ProjectDescription_Figure2.png"), width = 16,height = 15,unit = "in",res = 300)
fig1a + fig1b + fig1c + fig1d + fig1e + plot_layout(design = design)+ plot_annotation(tag_levels = 'A')&theme(plot.tag = element_text(size = 20, face = 'bold'))
dev.off()

```



## 4) Session info

```{r Session info, message = TRUE}

sessionInfo()

```
